# use faster csr indexing