@stegben 2016-12-24
Other candidate might be using `lil_matrix` since its row-iterating is fast, too. However, there're gonna be extra transforming time.